+++ tests/fixtures/v0.1.1_test.sh
@@ -1,0 +1,36 @@
+#! /bin/bash
+DIR=$(dirname $(readlink -f "$0"))
+
+################ Unit tests ################
+
+
+
+################ RUN shunit2 ################
+findShunit2()
+{
+    local curdir=$(dirname $(readlink -f "$1"))
+    while [ ! -e "${curdir}/lib/shunit2" ] && [ "${curdir}" != "/" ]; do
+        curdir=$(dirname ${curdir})
+    done
+
+    if [ "${curdir}" == "/" ]; then
+        echo "Error Shunit2 not found !" >&2
+        exit 1
+    fi
+
+    echo "${curdir}/lib/shunit2"
+}
+
+exitOnError()
+{
+    echo "$2" >&2
+    exit $1
+}
+#
+path=$(findShunit2 "$0")
+code=$?
+if [ ${code} -ne 0 ]; then
+    exitOnError ${code} "${path}"
+fi
+. "${path}"/source/2.1/src/shunit2
+#