+++ tests/fixtures/vcurrent_test.sh
@@ -1,0 +1,83 @@
+#! /bin/bash
+
+################ Utilities #################
+# get the full name of this script
+me()
+{
+    echo "$(readlink -f $BASH_SOURCE)"
+}
+
+# get the path of parent directory of this script
+mydir()
+{
+    echo "$(dirname $(me))"
+}
+
+# get the path of the project main directory
+maindir()
+{
+    local curdir="$(mydir)"
+
+    while \
+        [ ! -e "${curdir}/lib" ] && \
+        [ ! -e "${curdir}/src" ] && \
+        [ ! -e "${curdir}/tests" ] && \
+        [ "${curdir}" != "/" ]; do
+
+        curdir="$(dirname ${curdir})"
+    done
+
+    echo "${curdir}"
+}
+
+# get the path of the source directory
+srcdir()
+{
+    echo "$(maindir)/src"
+}
+
+# get the path of the test directory
+testdir()
+{
+    echo "$(maindir)/tests"
+}
+
+# get the path of the main lib directory
+libdir()
+{
+    echo "$(maindir)/lib"
+}
+
+################ Unit tests ################
+
+
+
+################ RUN shunit2 ################
+findShunit2()
+{
+    local curdir=$(dirname $(readlink -f "$1"))
+    while [ ! -e "${curdir}/lib/shunit2" ] && [ "${curdir}" != "/" ]; do
+        curdir=$(dirname ${curdir})
+    done
+
+    if [ "${curdir}" == "/" ]; then
+        echo "Error Shunit2 not found !" >&2
+        exit 1
+    fi
+
+    echo "${curdir}/lib/shunit2"
+}
+
+exitOnError()
+{
+    echo "$2" >&2
+    exit $1
+}
+#
+path=$(findShunit2 "$BASH_SOURCE")
+code=$?
+if [ ${code} -ne 0 ]; then
+    exitOnError ${code} "${path}"
+fi
+. "${path}"/source/2.1/src/shunit2
+## version: {{version}}